langgraph/experiments/agent-service/pyproject.toml
[project]
name = "agent-service"
version = "0.1.0"
description = ""
authors = [
    {name = "jhlee",email = "39754095+everError@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langchain-google-genai (>=2.1.9,<3.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langgraph (>=0.6.5,<0.7.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-openai (>=0.3.30,<0.4.0)"
]

[tool.poetry]
packages = [{include = "agent_service", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
